test: 환경변수에 따라 테스트 클래스 분리
